Ajuste en tabla, migración y controlador en confirmacion

+++ database/migrations/2019_03_02_175656_create_people_table.php
@@ -14,7 +14,7 @@ class CreatePeopleTable extends Migration
     public function up()
     {
         Schema::create('people', function (Blueprint $table) {
-            $table->bigInteger('id');
+            $table->increments('id');
             $table->string('name');
             $table->string('lastName');
             $table->string('secondLastName');
